\name{mrpt-package}
\alias{mrpt-package}
\alias{mrpt}
\docType{package}
\title{Mark-Recapture Distance Sampling for Lines and Points}
\description{Fits single and double observer point and line transect sampling models}
\details{
\tabular{ll}{
Package: \tab mrpt\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-04-19\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{Jeff Laake

Maintainer: <jeff.laake@noaa.gov>
}
\references{
Borchers, D. L., J. L. Laake, C. Southwell and C. G. M. Paxton (2006). "Accommodating unmodeled heterogeneity in double-observer distance sampling surveys." Biometrics 62(2): 372-378.

Buckland, S. T., J. Laake and D. Borchers (2010). "Double observer line transect methods: levels of independence." Biometrics 66: 169-177.

Laake and Collier. In Press. "Mark-Recapture Distance Sampling for Points"

Laake, J., M. J. Dawson and J. Hone (2008). "Visibility bias in aerial survey: mark-recapture, line-transect or both?" Wildlife Research 35(4): 299-309.

Laake, J. (1999). Distance sampling with independent observers: Reducing bias from heterogeneity by weakening the conditional independence assumption. Marine Mammal Survey and Assessment Methods. Rotterdam, A A BALKEMA: 137-148.

}
